+++ src/styles/index.css
@@ -24,3 +24,4 @@
 @import "./document-renamer.css";
 @import "./my-factoids.css";
 @import "./help.css";
+@import "./task-list.css";
