+++ src/styles/task-list.css
@@ -1,0 +1,8 @@
+.task-list {
+  position: absolute;
+  top: -83px;
+  left: 40px;
+  height: 100vh;
+  width: 20em;
+  background-color: red;
+}
